refactor: split Document utilities by concern
