+++ cutter2/Document/Document+Utilities.swift
@@ -11,275 +11,6 @@ import AVFoundation
 import os.log
 
 /* ============================================ */
-// MARK: - Actor isolation
-/* ============================================ */
-
-/// A simple thread-safe container for holding mutable results captured by @Sendable closures.
-///
-/// This class uses `@unchecked Sendable` because:
-/// - Access is synchronized via a DispatchQueue in `performAsync`
-/// - The value is only written once and read once
-/// - No concurrent access occurs during normal operation
-/// - The DispatchQueue provides the necessary memory barrier
-private final class SendableBox<T> {
-    var value: T
-    init(_ value: T) { self.value = value }
-}
-
-extension SendableBox: @unchecked Sendable {}
-
-extension Document {
-    
-    /// Executes an asynchronous, throwing operation synchronously on a detached task.
-    ///
-    /// This method bridges async/await operations to synchronous AppKit document APIs.
-    /// It is designed to be called from `nonisolated` contexts (specifically `Document.write(...)`).
-    ///
-    /// **Design Rationale:**
-    /// - `Task.detached` is used because:
-    ///   - Must be called from a `nonisolated` context (no parent task to inherit from)
-    ///   - AppKit's `canAsynchronouslyWrite` already executes `write()` on a background queue
-    ///   - No task priority inheritance is needed (AppKit manages thread priority)
-    /// - Semaphore waiting is used because:
-    ///   - Must block until async operation completes (AppKit document save requires synchronous return)
-    ///   - DispatchQueue provides thread-safe result passing
-    /// - Works in conjunction with `canAsynchronouslyWrite() -> Bool { true }`
-    ///
-    /// - Parameter block: A closure that performs asynchronous work and may throw.
-    /// - Returns: The result produced by the closure.
-    /// - Throws: An error thrown by the closure.
-    /// - Warning: This blocks the current thread. Do not call from the main thread.
-    nonisolated func performAsync<T: Sendable>(_ block: @Sendable @escaping () async throws -> T) throws -> T {
-        let semaphore = DispatchSemaphore(value: 0)
-        let lock = DispatchQueue(label: "ResultLock")
-        let resultBox = SendableBox<Result<T, Error>?>(nil)
-        Task.detached(priority: .userInitiated) { @Sendable in
-            let taskResult: Result<T, Error>
-            do {
-                taskResult = .success(try await block())
-            } catch {
-                taskResult = .failure(error)
-            }
-            lock.sync {
-                resultBox.value = taskResult
-            }
-            semaphore.signal()
-        }
-        semaphore.wait()
-        return try lock.sync { try resultBox.value!.get() }
-    }
-    
-    /// Executes an asynchronous, non-throwing operation synchronously on a detached task.
-    ///
-    /// This is the non-throwing variant of `performAsync`. See the throwing version for detailed rationale.
-    ///
-    /// - Parameter block: A closure that performs asynchronous work.
-    /// - Returns: The result produced by the closure.
-    /// - Warning: This blocks the current thread. Do not call from the main thread.
-    nonisolated func performAsync<T: Sendable>(_ block: @Sendable @escaping () async -> T) -> T {
-        let semaphore = DispatchSemaphore(value: 0)
-        let lock = DispatchQueue(label: "ResultLock")
-        let resultBox = SendableBox<T?>(nil)
-        Task.detached(priority: .userInitiated) { @Sendable in
-            let taskResult = await block()
-            lock.sync {
-                resultBox.value = taskResult
-            }
-            semaphore.signal()
-        }
-        semaphore.wait()
-        return lock.sync { resultBox.value! }
-    }
-    
-    /// Runs a throwing `@MainActor`-isolated closure synchronously.
-    /// - Parameter block: A closure isolated to the main actor that may throw an error.
-    /// - Returns: The result of the closure's operation.
-    /// - Throws: Any error thrown by the closure.
-    /// - Warning: Blocks the calling thread if not already on the main thread, potentially causing UI freezes.
-    nonisolated func performSyncOnMainActor<T: Sendable>(_ block: @MainActor () throws -> T) throws -> T {
-        return try ActorUtilities.performSyncOnMainActor(block)
-    }
-    
-    /// Runs a non-throwing `@MainActor`-isolated closure synchronously.
-    /// - Parameter block: A non-throwing closure isolated to the main actor.
-    /// - Returns: The result of the closure's operation.
-    /// - Warning: Blocks the calling thread if not already on the main thread, potentially causing UI freezes.
-    nonisolated func performSyncOnMainActor<T: Sendable>(_ block: @MainActor () -> T) -> T {
-        return ActorUtilities.performSyncOnMainActor(block)
-    }
-}
-
-/* ============================================ */
-// MARK: - Sheet control
-/* ============================================ */
-
-extension Document {
-    
-    /// Update progress with smooth animation
-    ///
-    /// This function implements exponential smoothing for smoother progress bar animation.
-    /// It also throttles UI updates to every 100ms to reduce unnecessary redraws.
-    ///
-    /// - Parameter progress: The raw progress value (0.0 to 1.0)
-    public func updateProgress(_ progress: Float) {
-        
-        // Use Low frequency update (throttle to 100ms)
-        let unit = NSEC_PER_MSEC * 100 // 100ms
-        let t: UInt64 = clock_gettime_nsec_np(CLOCK_REALTIME)
-        if lastUpdateAt == 0 {
-            // First call: initialize timestamp and progress
-            lastUpdateAt = t
-            lastReportedProgress = 0.0
-        } else {
-            if t > lastUpdateAt {
-                // Normal case: time moved forward
-                if (t - lastUpdateAt) > unit {
-                    // Sufficient time passed: update timestamp
-                    lastUpdateAt += unit
-                } else {
-                    // Too soon: skip this update
-                    return
-                }
-            } else {
-                // Edge case: clock went backwards, reset
-                lastUpdateAt = t
-            }
-        }
-        
-        // Apply exponential smoothing for smoother animation
-        // Formula: smoothed = α * new + (1 - α) * old
-        // α = 0.3 provides good balance between responsiveness and smoothness
-        let smoothingFactor: Float = 0.3
-        let smoothedProgress = lastReportedProgress + smoothingFactor * (progress - lastReportedProgress)
-        lastReportedProgress = smoothedProgress
-        
-        // Update UI in main queue
-        Task { @MainActor in
-            
-            guard let alert = self.alert else { return }
-            guard smoothedProgress.isNormal else { return }
-            
-            // Update progress indicator (visual feedback)
-            if let indicator = self.progressIndicator {
-                indicator.doubleValue = Double(smoothedProgress * 100.0)
-            }
-            
-            // Update text (percentage)
-            let format = NSLocalizedString("progress.format.percent", comment: "Progress percentage format")
-            alert.informativeText = String(format: format, Int(smoothedProgress * 100))
-        }
-    }
-    
-    /// Show busy modalSheet
-    ///
-    /// Displays a modal sheet with progress information, a visual progress indicator,
-    /// and an optional Cancel button.
-    /// When the user clicks Cancel, the operation is cancelled via NSProgress and MovieMutator.
-    ///
-    /// - Parameters:
-    ///   - message: The main message to display (defaults to "Processing...")
-    ///   - info: Additional information text (defaults to "Hold on seconds...")
-    public func showBusySheet(_ message: String?, _ info: String?) {
-        
-        // Reset progress state for new operation
-        lastUpdateAt = 0
-        lastReportedProgress = 0.0
-        
-        Task { @MainActor in
-            
-            guard let window = self.window else { return }
-            
-            // Create progress indicator for visual feedback
-            let progressIndicator = NSProgressIndicator()
-            progressIndicator.style = .bar
-            progressIndicator.isIndeterminate = false
-            progressIndicator.minValue = 0.0
-            progressIndicator.maxValue = 100.0
-            progressIndicator.doubleValue = 0.0
-            progressIndicator.controlSize = .regular
-            
-            // Set frame for progress indicator (width: 300pt, height: 20pt)
-            progressIndicator.frame = NSRect(x: 0, y: 0, width: 300, height: 20)
-            
-            // Store reference for updates
-            self.progressIndicator = progressIndicator
-            
-            // Create alert
-            let alert: NSAlert = NSAlert()
-            let defaultTitle = NSLocalizedString("progress.default.title", comment: "Default title for progress dialog")
-            let defaultMessage = NSLocalizedString("progress.default.message", comment: "Default message for progress dialog")
-            alert.messageText = message ?? defaultTitle
-            alert.informativeText = info ?? defaultMessage
-            alert.alertStyle = .informational
-            
-            // Add progress indicator as accessory view
-            alert.accessoryView = progressIndicator
-            
-            alert.addButton(withTitle: NSLocalizedString("ui.button.cancel", comment: "Cancel button for canceling operations"))
-            let handler: (NSApplication.ModalResponse) -> Void = { @Sendable [weak self] (response) in // @escaping
-                guard let self else { return }
-                if response == .alertFirstButtonReturn {
-                    // User clicked Cancel - the cancellationHandler will call mutator.cancel()
-                    performSyncOnMainActor {
-                        self.saveProgress?.cancel()
-                    }
-                }
-            }
-            alert.beginSheetModal(for: window, completionHandler: handler)
-            
-            // Keep NSAlert object for later update
-            self.alert = alert
-        }
-    }
-    
-    /// Hide busy modalSheet
-    public func hideBusySheet() {
-        
-        Task { @MainActor in
-            
-            guard let window = self.window else { return }
-            guard let alert = self.alert else { return }
-            
-            window.endSheet(alert.window)
-            
-            // Release NSAlert and progress indicator
-            self.alert = nil
-            self.progressIndicator = nil
-        }
-    }
-    
-    /// Present ErrorSheet asynchronously
-    public func showErrorSheet(_ error: Error) {
-        
-        // Don't use NSDocument default error handling
-        guard let window = self.window else { preconditionFailure("Unexpected nil window detected.") }
-        Task { @MainActor in
-            
-            let alert = NSAlert(error: error)
-            let err :NSError = error as NSError
-            var text :String? = nil
-            let userInfo: [String:Any] = err.userInfo // Can be empty dictionary
-            if userInfo.count > 0 {
-                let keys = userInfo.keys
-                if keys.contains(NSUnderlyingErrorKey) || keys.contains(NSDebugDescriptionErrorKey) {
-                    text = err.description
-                } else if keys.contains(NSLocalizedFailureErrorKey) {
-                    text = userInfo[NSLocalizedFailureErrorKey] as? String
-                } else if keys.contains(NSLocalizedDescriptionKey) {
-                    text = userInfo[NSLocalizedDescriptionKey] as? String
-                } else if keys.contains(NSLocalizedFailureReasonErrorKey) {
-                    text = userInfo[NSLocalizedFailureReasonErrorKey] as? String
-                }
-            }
-            if let text = text {
-                alert.informativeText = text
-            }
-            alert.beginSheetModal(for: window, completionHandler: nil)
-        }
-    }
-}
-
-/* ============================================ */
 // MARK: - Misc utilities
 /* ============================================ */
 
@@ -482,345 +213,5 @@ extension Document {
     func updateRate(_ player: AVPlayer, _ newRate: Float) {
         guard player.rate != newRate else { return }
         player.rate = newRate
-    }
-}
-
-/* ============================================ */
-// MARK: - Observer
-/* ============================================ */
-
-extension Document {
-    
-    /// Add AVPlayer properties observer
-    private func addPlayerObserver() {
-        
-        guard let player = self.player else { return }
-        
-        player.addObserver(self,
-                           forKeyPath: #keyPath(AVPlayer.status),
-                           options: [.old, .new],
-                           context: &(self.kvoContext))
-        player.addObserver(self,
-                           forKeyPath: #keyPath(AVPlayer.rate),
-                           options: [.old, .new],
-                           context: &(self.kvoContext))
-    }
-    
-    /// Remove AVPlayer properties observer
-    private func removePlayerObserver() {
-        
-        guard let player = self.player else { return }
-        
-        player.removeObserver(self,
-                              forKeyPath: #keyPath(AVPlayer.status),
-                              context: &(self.kvoContext))
-        player.removeObserver(self,
-                              forKeyPath: #keyPath(AVPlayer.rate),
-                              context: &(self.kvoContext))
-    }
-    
-    /// compare KVO context address as UInt
-    @MainActor func checkKVOContext(_ contextAddress: UInt) -> Bool {
-        return withUnsafePointer(to: &self.kvoContext) { kvoPointer in
-            let kvoAddress = UInt(bitPattern: kvoPointer)
-            return (contextAddress == kvoAddress)
-        }
-    }
-    
-    // NSKeyValueObserving protocol - observeValue(forKeyPath:of:change:context:)
-    override nonisolated func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey:Any]?,
-                                           context: UnsafeMutableRawPointer?) {
-        
-        guard
-            let context = context, let object = object as? AVPlayer, let keyPath = keyPath, let change = change
-        else {
-            super.observeValue(forKeyPath: keyPath,
-                               of: object,
-                               change: change,
-                               context: context)
-            return
-        }
-        
-        let contextAddress = UInt(bitPattern: context) // Cast UnsafeMutableRawPointer to UInt for actor isolation
-        let (objectIsPlayer, keyPathIsAVPlayerStatus, keyPathIsAVPlayerRate) = performSyncOnMainActor {
-            let contextMatch: Bool = checkKVOContext(contextAddress)
-            let objectIsPlayer: Bool = (object === self.player)
-            let keyPathIsAVPlayerStatus: Bool = (keyPath == #keyPath(AVPlayer.status))
-            let keyPathIsAVPlayerRate: Bool = (keyPath == #keyPath(AVPlayer.rate))
-            return (contextMatch && objectIsPlayer, keyPathIsAVPlayerStatus, keyPathIsAVPlayerRate)
-        }
-        
-        if objectIsPlayer && keyPathIsAVPlayerStatus {
-            
-            // Force redraw when AVPlayer.status is updated
-            let newStatus = change[.newKey] as! NSNumber
-            if newStatus.intValue == AVPlayer.Status.readyToPlay.rawValue {
-                // Seek and refresh View
-                performSyncOnMainActor {
-                    guard let mutator = self.movieMutator else { return }
-                    let time = mutator.insertionTime
-                    let range = mutator.selectedTimeRange
-                    updateGUI(time, range, false)
-                }
-            } else if newStatus.intValue == AVPlayer.Status.failed.rawValue {
-                //
-                LoggingSystem.ui.error("AVPlayerStatus.failed detected")
-            }
-            return
-        } else if objectIsPlayer && keyPathIsAVPlayerRate {
-            
-            // Check special case: movie play reached at end of movie
-            let oldRate = change[.oldKey] as! NSNumber
-            let newRate = change[.newKey] as! NSNumber
-            if oldRate.floatValue > 0.0 && newRate.floatValue == 0.0 {
-                // Movie stopped
-                performSyncOnMainActor {
-                    guard let player = self.player else { return }
-                    guard let mutator = self.movieMutator else { return }
-                    
-                    // Check if it is tail of movie
-                    let current = player.currentTime()
-                    let duration = mutator.movieDuration()
-                    let selection = mutator.selectedTimeRange
-                    if current == duration {
-                        // Force-refresh GUI at the end of movie
-                        updateTimeline(current, range: selection)
-                    }
-                }
-                LoggingSystem.ui.debug("Movie playback stopped")
-            }
-            if oldRate.floatValue == 0.0 && newRate.floatValue > 0.0 {
-                LoggingSystem.ui.debug("Movie playback started (forward)")
-            }
-            if oldRate.floatValue == 0.0 && newRate.floatValue < 0.0 {
-                LoggingSystem.ui.debug("Movie playback started (backward)")
-            }
-            if oldRate.floatValue == newRate.floatValue {
-                LoggingSystem.ui.notice("No rate change detected - needs investigation")
-            }
-            return
-        } else {
-            super.observeValue(forKeyPath: keyPath,
-                               of: object,
-                               change: change,
-                               context: context)
-        }
-    }
-    
-    /// Register observer for movie mutation
-    public func addMutationObserver() {
-        
-        let handler: @Sendable (Notification) -> Void = {[weak self] (notification) in // @escaping
-            
-            guard let self else { preconditionFailure("Unexpected nil self detected.") }
-            guard
-                let mutator = performSyncOnMainActor({ self.movieMutator }),
-                let object = notification.object as? MovieMutator,
-                mutator == object
-            else { return }
-            
-            #if DEBUG
-            let displayName = performSyncOnMainActor({ self.displayName ?? "unknown" })
-            LoggingSystem.document.debug("Received .movieWasMutated notification: \(displayName)")
-            #endif
-            
-            // extract CMTime/CMTimeRange from userInfo
-            guard
-                let userInfo = notification.userInfo,
-                let timeValue = userInfo[timeValueInfoKey] as? NSValue,
-                let timeRangeValue = userInfo[timeRangeValueInfoKey] as? NSValue
-            else { return }
-            
-            let time: CMTime = timeValue.timeValue
-            let timeRange: CMTimeRange = timeRangeValue.timeRangeValue
-            performSyncOnMainActor {
-                updateGUI(time, timeRange, true)
-            }
-        }
-        do {
-            guard let mutator = self.movieMutator else { return }
-            let center = NotificationCenter.default
-            var observer: NSObjectProtocol? = nil
-            observer = center.addObserver(forName: .movieWasMutated,
-                                          object: mutator,
-                                          queue: OperationQueue.main,
-                                          using: handler)
-            self.mutationObserver = observer
-        }
-    }
-    
-    /// Unregister observer for movie mutation
-    public func removeMutationObserver() {
-        
-        do {
-            guard let mutator = self.movieMutator else { return }
-            guard let observer = self.mutationObserver else { return }
-            let center = NotificationCenter.default
-            center.removeObserver(observer,
-                                  name: .movieWasMutated,
-                                  object: mutator)
-            self.mutationObserver = nil
-        }
-    }
-    
-    /// Unregister all undo record for current MovieMutator object
-    public func removeAllUndoRecords() {
-        
-        guard let mutator = self.movieMutator else { return }
-        self.undoManagerWrapper.removeAllActions(withTarget: mutator)
-    }
-}
-
-/* ============================================ */
-// MARK: -  Position control
-/* ============================================ */
-
-extension Document {
-    
-    /// Move either start/end marker at current marker (nearest marker do sync)
-    public func syncSelection(_ current: CMTime) {
-        
-        guard let mutator = self.movieMutator else { return }
-        let selection: CMTimeRange = mutator.selectedTimeRange
-        let start: CMTime = selection.start
-        let end: CMTime = selection.end
-        
-        let halfDuration: CMTime = CMTimeMultiplyByRatio(selection.duration, multiplier: 1, divisor: 2)
-        let centerOfRange: CMTime = start + halfDuration
-        let t1: CMTime = (current < centerOfRange) ? current : start
-        let t2: CMTime = (current > centerOfRange) ? current : end
-        let newSelection: CMTimeRange = CMTimeRangeFromTimeToTime(start: t1, end: t2)
-        mutator.selectedTimeRange = newSelection
-    }
-    
-    /// Move either Or both start/end marker to current marker
-    public func resetSelection(_ newTime: CMTime, _ resetStart: Bool, _ resetEnd: Bool) {
-        
-        guard let mutator = self.movieMutator else { return }
-        let selection: CMTimeRange = mutator.selectedTimeRange
-        let start: CMTime = selection.start
-        let end: CMTime = selection.end
-        
-        let sFlag: Bool = (resetEnd && newTime < start) ? true : resetStart
-        let eFlag: Bool = (resetStart && newTime > end) ? true : resetEnd
-        if sFlag || eFlag {
-            let t1: CMTime = sFlag ? newTime : start
-            let t2: CMTime = eFlag ? newTime : end
-            let newSelection: CMTimeRange = CMTimeRangeFromTimeToTime(start: t1, end: t2)
-            mutator.selectedTimeRange = newSelection
-        }
-    }
-    
-    /// Check if it is head of movie
-    public func checkHeadOfMovie() -> Bool {
-        
-        guard let player = self.player else { return false }
-        
-        // NOTE: Return false if player is not paused.
-        if player.rate != 0.0 { return false }
-        
-        let current = player.currentTime()
-        if current == CMTime.zero {
-            return true
-        }
-        return false
-    }
-    
-    private func debugTrackRange(_ range: CMTimeRange, _ current: CMTime, _ endOfRange: Bool) {
-        #if DEBUG
-        guard let mutator = self.movieMutator else { return }
-        let rangeStart = String(format: "%4.3f", range.start.seconds)
-        let rangeEnd = String(format: "%4.3f", range.end.seconds)
-        let currentStr = String(format: "%4.3f", current.seconds)
-        let insertionStr = String(format: "%4.3f", mutator.insertionTime.seconds)
-        let timeDiff = current == mutator.insertionTime ? "" : " (diff)"
-        let contains = (range.start <= current && current <= range.end)
-        let endNote = endOfRange ? " - End of Movie" : ""
-        
-        LoggingSystem.video.debug("Track range: [\(rangeStart), \(rangeEnd)], current: \(currentStr), insertion: \(insertionStr)\(timeDiff), containsTime: \(contains)\(endNote)")
-        #endif
-    }
-    
-    /// Check if it is tail of movie
-    public func checkTailOfMovie() -> Bool {
-        
-        guard let mutator = self.movieMutator else { return false }
-        guard let player = self.player else { return false }
-        
-        // NOTE: Return false if player is not paused.
-        if player.rate != 0.0 { return false }
-        
-        let current = player.currentTime()
-        let duration: CMTime = mutator.movieDuration()
-        
-        // validate cached range value
-        if let range = cachedLastSampleRange, range.start <= current, current <= range.end {
-            // use cached result
-            // debugTrackRange(range, current, true)
-            return cachedWithinLastSampleRange
-        } else {
-            // reset cache
-            cachedTime = current
-            cachedWithinLastSampleRange = false
-            cachedLastSampleRange = CMTimeRange.invalid
-            
-            if let info = mutator.presentationInfoAtTime(current) {
-                let endOfRange: Bool = info.timeRange.end == duration
-                if endOfRange {
-                    cachedTime = current
-                    cachedWithinLastSampleRange = true
-                    cachedLastSampleRange = info.timeRange
-                }
-                // debugTrackRange(info.timeRange, current, endOfRange)
-            }
-            return cachedWithinLastSampleRange
-        }
-    }
-    
-    /// Snap to grid - Adjust Timeline resolution
-    public func quantize(_ position: Float64) -> CMTime {
-        
-        guard let mutator = self.movieMutator else { return CMTime.zero }
-        let position: Float64 = min(max(position, 0.0), 1.0)
-        if let info = mutator.presentationInfoAtPosition(position) {
-            let ratio: Float64 = (position - info.startPosition) / (info.endPosition - info.startPosition)
-            return (ratio < 0.5) ? info.timeRange.start : info.timeRange.end
-        } else {
-            return CMTimeMultiplyByFloat64(mutator.movieDuration(), multiplier: position)
-        }
-    }
-}
-
-/* ============================================ */
-// MARK: -  Movie Reference utility
-/* ============================================ */
-
-extension Document {
-    
-    public func validateIfSelfContained(for url: URL) -> Bool {
-        
-        /*
-         If a movie refers to one file path only and it is same as the movie's filePath,
-         - the URL is the only one source of the movie
-         - movie file is self-containd - no referencing track is included
-         
-         In case of in-memory movie (no-file-backed) it should be a reference movie.
-         In case of multiple url found it should be a reference movie.
-         */
-        let refURLs: [URL] = self.movieMutator?.queryMediaDataURLs() ?? []
-        if refURLs.count == 1 && refURLs[0] == url {
-            return true
-        }
-        //
-        if refURLs.count < 1 {
-            LoggingSystem.video.error("Unable to get track reference URLs")
-        }
-        if refURLs.count == 1 {
-            LoggingSystem.video.info("Different track reference URL found")
-        }
-        if refURLs.count > 1 {
-            LoggingSystem.video.info("Multiple track reference URLs found")
-        }
-        return false
     }
 }
